Rename StartupModal folder to Modal
