+++ src/components/Modal/index.js
@@ -1,0 +1,5 @@
+import Modal from "./Modal";
+import StartupModal from "./StartupModal";
+
+export { StartupModal };
+export default Modal;
