--- src/components/StartupModal/index.js
@@ -1,3 +1,0 @@
-import StartupModal from "./StartupModal";
-
-export default StartupModal;
