+++ src/components/Table/Table.tsx
@@ -16,7 +16,7 @@ import Controls from "../Controls";
 import MainPot from "./MainPot";
 import Game from "../Game";
 import Connections from "./Connections";
-import StartupModal from "../StartupModal";
+import { StartupModal } from "../Modal";
 import DeveloperMode from "../DeveloperMode";
 import LogBox from "../LogBox";
